ConstructiveMcdc

Documentation

(<?>) :: Maybe a -> e -> Either e a

A useful Maybe to Eitehr e mapper, using the given constant value when a Left is needed.

data JoinError a

Possible errors from joining ValueAssignments

Constructors

NoSingleElement

ValueConflict (Set a)

NoMcdcCoverage (ValueAssignment a)

joinAnd :: Ord a => ValueAssignment a -> ValueAssignment a -> Either (JoinError a) (ValueAssignment a)

Join 2 ValueAssignments using and logic: Take all TruthVectors from a and merge them with a true value from b (because True is the identity of and) Take all TruthVectors from b and merge them with a true value from a. Reassemble the items into their new true and false value sets and return the new TruthVector.

joinAnd2 :: Ord a => ValueAssignment a -> ValueAssignment a -> Either (JoinError a) (ValueAssignment a)

joinOr :: Ord a => ValueAssignment a -> ValueAssignment a -> Either (JoinError a) (ValueAssignment a)

Same as joinAnd, but for or. Implemented using the equivalence "A || B == !(!A && !B)"

joinOr2 :: Ord a => ValueAssignment a -> ValueAssignment a -> Either (JoinError a) (ValueAssignment a)

findMcdc :: Ord a => BoolExp a -> Either (JoinError a) (ValueAssignment a)

BoolExp evaluator which gives a MCDC minimal set for a given expression.