Expressions

Documentation

data BoolExp a

A basic boolean expression tree.

Constructors

Var a

And (BoolExp a) (BoolExp a)

Or (BoolExp a) (BoolExp a)

Not (BoolExp a)

data Evaluator a b m

Functions for how to deal with a 'BoolExp a', in some monad m.

Constructors

Evaluator

Fields

evalDoNot :: b -> m b

evalDoAnd :: b -> b -> m b

evalDoOr :: b -> b -> m b

evalLookup :: a -> m b

pureEvaluator :: (b -> b) -> (b -> b -> b) -> (b -> b -> b) -> (a -> b) -> Evaluator a b Identity

Smart constructor which embeds the pure functions given in the Identity Monad to make an Evaluator

monadEval :: Monad m => Evaluator a b m -> BoolExp a -> m b

Evaluate an expression using the given evaluator in a Monad.

generalEval :: Evaluator a b Identity -> BoolExp a -> b

Evaluate and expression using the Identity monad and unwrap.

allVars :: Ord a => BoolExp a -> Set a

find all variables in expression

eval :: (a -> Bool) -> BoolExp a -> Bool

Evaluate the value of an expression using the given name to boolean value mapping.

evalSet :: Ord a => Set a -> BoolExp a -> Bool

Evaluate the value of an expression using the given set of names of true variables.