McdcCheck

Documentation

hasMcdcPairFor :: Ord a => ValueAssignment a -> a -> Bool

small MCDC check -- does the given ValueAssignment have an entry in trues containing target, and an entry in falses containing target, such that the vectors only differ in the value assigned to @target.

hasMcdcCoverage :: Ord a => ValueAssignment a -> Bool

large MCDC check -- does the valueassignment have good pairs for all variables?

isNPlus1Coverage :: Ord a => ValueAssignment a -> Bool

Is the size of the ValueAssignment equal to one more than the number of variables?

isMinimalMcdcCoverage :: Ord a => ValueAssignment a -> Bool

Is the ValueAssignment correct and minimal?