TruthVectors

Documentation

data TruthVector a

A TruthVector of a is a pair of sets falses and trues. An element is false if it is in the first set, true if it is in the second. If it is in neither the value is not (yet) defined and if in both then the vector has a conflict.

Constructors

TruthVector

Fields

falses, trues :: Set a

empty :: TruthVector a

The empty TruthVector

insert :: Ord a => a -> Bool -> TruthVector a -> TruthVector a

Insert a new variable with the given value.

singletonTrue :: Ord a => a -> TruthVector a

Take a single value and mark it true in a new TruthVector

singletonFalse :: Ord a => a -> TruthVector a

Take a single value and mark it false in a new TruthVector

fromLists :: Ord a => [a] -> [a] -> TruthVector a

Create a TruthVector from a list of false items and a list of true items.

fromFuncAndSet :: Ord a => (a -> Bool) -> Set a -> TruthVector a

Create a TruthVector from a list of items and a function to assign truth to them.

invert :: TruthVector a -> TruthVector a

Swap the falses and trues sets, making everything that was false true and vice versa.

elements :: Ord a => TruthVector a -> Set a

Return a set of all elements defined in the TruthVector

delete :: Ord a => a -> TruthVector a -> TruthVector a

Unmark an item as true and false -- remove it from both sets.

lookup :: Ord a => a -> TruthVector a -> Maybe Bool

Determine the value of a given element. Returns Nothing for conflicted or undefined.

merge :: Ord a => TruthVector a -> TruthVector a -> TruthVector a

Merges 2 TruthVectors naively. The resulting vector may have conflicts

contradictions :: Ord a => TruthVector a -> Set a

Find all contradictions in the TruthVector.

safeMerge :: Ord a => TruthVector a -> TruthVector a -> Either (Set a) (TruthVector a)

Try to merge 2 TruthVectors. If the resulting TruthVector is conflicted then return Nothing, otherwise return the result.

joinVals :: Ord a => Set (TruthVector a) -> TruthVector a -> Either (Set a) (Set (TruthVector a))

Join each element of the set as with b. If any of the resulting TruthVectors are conflicted then return a Left value containing the first conflicts found. Otherwise return the new TruthVectors.

joinVals2 :: Ord a => Set (TruthVector a) -> Set (TruthVector a) -> Set (TruthVector a)

Join each element of the set as with each element of the set bs. If any of the resulting TruthVectors are conflicted then return a Left value containing the first conflicts found. Otherwise return the new TruthVectors.

prettyPrint :: Ord a => (a -> String) -> (a -> String) -> TruthVector a -> String

Function to help print out a TruthVector - given a false-printing function and a true printing function and a TruthVector, produce a String. Variables are output in order Example: prettyPrint (:[]) ((:[]) . Char.toUpper) tv == "aBc"

prettyPrint2 :: (a -> String) -> TruthVector a -> String

Function to help print out a TruthVector -- given an element-printing function, first print all false elements, then all true elements. Example: prettyPrint (:[]) ((:[]) . Char.toUpper) tv == "FALSE: ac TRUE: b"