ValueAssignment

Documentation

type ValueAssignment a = TruthVector (TruthVector a)

A ValueAssignment is a set of TruthVectors, but separated out into those marked true and false. That is, it is a TruthVector of TruthVectors.

singleton :: Ord a => a -> ValueAssignment a

Returns a vector where the single value @a is marked false in the false set, and true in the true set.

allVarsInAssignment :: Ord a => ValueAssignment a -> [a]

Get all variables defined in all truthvectors.